requirements.txt
arrow
Django
qrcode
djangorestframework
djangorestframework-api-key
djangorestframework-datatables
git+https://github.com/phoebebright/django-docserve
django-anymail[mailgun]
django-tinymce
psycopg2-binary
boto3
django-storages
cryptography
django-sites
git+https://github.com/phoebebright/django-users@skorie_users
nanoid
django_countries
django-phonenumber-field[phonenumbers]
git+https://github.com/phoebebright/django-keycloak-admin
Django
twilio
nanoid
qrcode
djangorestframework
djangorestframework-api-key
djangorestframework-datatables
python-jose
git+https://github.com/phoebebright/django-docserve
django_countries
email-validator
pycountry

git+https://github.com/phoebebright/django-yamlfield
django-timezone-field
django_countries
nanoid
django-filter
Pillow

# this is in users and should not be...
zammad_py
